music_recommendation_app/requirements.txt
django
djangorestframework
django-cors-headers
mediapipe
tensorflow
pandas